Reorganize directories - create auth/routes dir
