+++ packages/neoBackend/src/auth/routes/index.js
@@ -1,0 +1,2 @@
+export { authGitHub } from './authGitHub';
+export { authMe } from './authMe';
